Decompose r_demo
