+++ prboom2/src/dsda/analysis.c
@@ -16,9 +16,9 @@
 //
 
 #include "doomstat.h"
-#include "r_demo.h"
 
 #include "dsda/excmd.h"
+#include "dsda/exdemo.h"
 #include "dsda/settings.h"
 
 #include "analysis.h"
@@ -84,7 +84,7 @@ void dsda_WriteAnalysis(void) {
   }
 
   category = dsda_DetectCategory();
-  is_signed = R_DemoEx_IsSigned();
+  is_signed = dsda_IsExDemoSigned();
 
   fprintf(fstream, "skill %d\n", gameskill + 1);
   fprintf(fstream, "nomonsters %d\n", dsda_nomo);
