+++ prboom2/src/dsda/skip.c
@@ -20,8 +20,8 @@
 #include "e6y.h"
 #include "i_main.h"
 #include "i_sound.h"
-#include "r_demo.h"
 #include "s_sound.h"
+#include "smooth.h"
 #include "v_video.h"
 #include "gl_struct.h"
