+++ prboom2/src/dsda/text_file.c
@@ -16,7 +16,6 @@
 //
 
 #include "doomstat.h"
-#include "r_demo.h"
 #include "lprintf.h"
 #include "e6y.h"
