+++ prboom2/src/e6y.h
@@ -37,7 +37,6 @@
 #include <stdarg.h>
 
 #include "hu_lib.h"
-#include "r_demo.h"
 
 #define HU_HUDADDX (HU_HUDX)
 #define HU_HUDADDY (HU_HUDY+(-1)*HU_GAPY)
